runtime: encapsulate rules and prioritisation in classifier
The runtime directly manipulates the rules encapsulated in the
classifier, breaking the abstraction. This moves the iteration logic
back into the classifier. The runtime can register "mapping" functions
with the classifier which will be applied in order to the rules in the
classifier when the rules_for_install() member function is called.
The rules_for_install() member also handles prioritisation of rules.
This allows custom classifiers to subclass and override prioritisation
logic.